+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2024-10-15
+
+- Added health-check-path, health-check-interval charm configurations.
+- Added backends-path charm configurations.
+- Added proxy-cache-valid charm configurations.
+
 ## 2024-10-07
 
 - Added a changelog for tracking user-relevant changes.